+++ cli/commands/subcommands/build.js
@@ -37,14 +37,25 @@ module.exports = function(program, config = {}) {
   // if config has minify option
   if (config.build.minify) {
     let options = { compress: { warnings: true } };
+    let defineOptions = { 'process.env': { 'NODE_ENV': JSON.stringify('production') } };
+
     let minifyServer = new webpack.optimize.UglifyJsPlugin(options);
     let minifyClient = new webpack.optimize.UglifyJsPlugin(options);
+
+    let defineServerProduction = new webpack.DefinePlugin(defineOptions);
+    let defineClientProduction = new webpack.DefinePlugin(defineOptions);
+
     // check plugins block
     serverConfig && !serverConfig.plugins ? serverConfig.plugins = [] : null;
     clientConfig && !clientConfig.plugins ? clientConfig.plugins = [] : null;
+
     // push minify plugins to config
     serverConfig.plugins.push(minifyServer);
     clientConfig.plugins.push(minifyClient);
+
+    // push define production env
+    serverConfig.plugins.push(defineServerProduction);
+    clientConfig.plugins.push(defineClientProduction);
   }
 
   // prepare separated webpack instances
@@ -58,6 +69,8 @@ module.exports = function(program, config = {}) {
   // prepare folders for compiling
   utils.exec(`rm -rf ${dir}/dist`, { cwd: cwd }, null, true);
   utils.exec(`mkdir ${dir}/dist`, { cwd: cwd }, null, true);
+  utils.exec(`mkdir ${dir}/dist/server`, { cwd: cwd }, null, true);
+  utils.exec(`mkdir ${dir}/dist/client`, { cwd: cwd }, null, true);
 
   let Server = new Promise((resolve, reject) => {
     serverCompiler.run((err, stat) => {
@@ -77,19 +90,29 @@ module.exports = function(program, config = {}) {
     });
   });
 
-  // status checked
-  Server.then(() => logger('SHIP: Webpack.Server:Success', 'green'))
-  Client.then(() => logger('SHIP: Webpack.Client:Success', 'green'))
-
   // copy node_modules
   let Copy = new Promise((resolve, reject) => {
     try {
-      utils.exec(`mkdir ${dir}/dist/server`, { cwd: cwd }, null, true);
-      utils.exec(`cp -r ${dir}/node_modules ${dir}/dist/server/node_modules/`, { cwd: cwd }, null, true);
+      utils.exec(`cp ${dir}/package.json ${dir}/dist/server`, { cwd: cwd }, null, true);
       resolve();
     } catch (err) {
       reject(err);
     }
+  });
+
+  // status checked
+  Server.then(() => logger('SHIP: Webpack.Server:Success', 'green'));
+  Client.then(() => logger('SHIP: Webpack.Client:Success', 'green'));
+
+  // lets install production dependencies
+  Copy.then(() => {
+    utils.exec(
+      utils.makeCommand(cwd, 'install --production', [], ''), // command
+        { cwd: `${dir}/dist/server` }, // options
+        null,         // no callback
+        false,        // no sync
+        true          // print stdout
+    );
   });
 
   // wrap steps to promise