+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+#### 1.0.13
+- Add Production configuration for webpack @see https://github.com/rambler-digital-solutions/rship/issues/17
+- Improve speed of building app
+
 #### 1.0.12
 - Remove unuseable lines of code
 - Update dependencies